Aliens and Machine blocks
Added the First Xeno to another planet, Magma monsters to Io

Began the process of making it so that machines (and eventually tools) use circuits and machine blocks associated with the tier they belong in.

Removed a bunch of machines that have their purpose done by earlier machines.

Added Text to a bunch of quests and moved stuff around.